=refactor and added new categorytest

+++ moduls/Badzohreh/Category/Http/Controllers/CategoryController.php
@@ -4,6 +4,7 @@ namespace Badzohreh\Category\Http\Controllers;
 
 use App\Http\Controllers\Controller;
 use Badzohreh\Category\Http\Requests\CategorySotreRequest;
+use Badzohreh\Category\Models\Category;
 use Badzohreh\Category\Repositories\CategoryRepo;
 use Badzohreh\Category\Responses\AjaxResponses;
 
@@ -18,15 +19,19 @@ class CategoryController extends Controller
 
     public function index()
     {
+        $this->authorize("manage",Category::class);
 
-        $user  =auth()->user();
-
+        /*todo: delete user varible*/
+        $user=auth()->user();
         $categories = $this->repo->all();
         return view("Categories::index", compact('categories'));
     }
 
     public function store(CategorySotreRequest $request)
     {
+
+        $this->authorize("manage",Category::class);
+
         $this->repo->store($request);
         return back();
     }
@@ -34,6 +39,8 @@ class CategoryController extends Controller
 
     public function edit($categoryId)
     {
+        $this->authorize("manage",Category::class);
+
         $category=$this->repo->findById($categoryId);
         $categories = $this->repo->allExpectId($categoryId);
         return view("Categories::edit", compact('category', 'categories'));
@@ -41,12 +48,15 @@ class CategoryController extends Controller
 
     public function update($categoryId, CategorySotreRequest $request)
     {
+        $this->authorize("manage",Category::class);
         $this->repo->update($categoryId,$request);
         return back();
     }
 
     public function destroy($categoryId)
     {
+        $this->authorize("manage",Category::class);
+
         $this->repo->delete($categoryId);
         return AjaxResponses::successResponses();
     }
